requirements.txt
torch==1.6.0
numpy>=1.14.5
PyYAML>=5.1
